move tests to files that parallel the locaiton of the functions
