+++ packages/shared/test/util/holdouts.test.ts
@@ -1,5 +1,6 @@
 import {
   coverageToHoldoutSize,
+  getHoldoutStage,
   holdoutSizeToCoverage,
   MAX_HOLDOUT_SIZE,
 } from "../../src/util/holdouts";
@@ -15,5 +16,66 @@ describe("holdout size / coverage conversion", () => {
     for (const size of [0.01, 0.03, 0.05, 0.07, 0.1, 0.15, 0.2, 0.33, 0.5]) {
       expect(coverageToHoldoutSize(holdoutSizeToCoverage(size))).toBe(size);
     }
+  });
+});
+
+describe("getHoldoutStage", () => {
+  it("returns draft for a draft holdout experiment", () => {
+    expect(getHoldoutStage({}, { status: "draft" })).toBe("draft");
+  });
+
+  it("returns draft when analysisStartDate is set", () => {
+    expect(
+      getHoldoutStage({ analysisStartDate: new Date() }, { status: "draft" }),
+    ).toBe("draft");
+  });
+
+  it("returns running while the holdout has not entered its analysis period", () => {
+    expect(getHoldoutStage({}, { status: "running" })).toBe("running");
+  });
+
+  it("returns analysis-period once analysisStartDate is set", () => {
+    expect(
+      getHoldoutStage({ analysisStartDate: new Date() }, { status: "running" }),
+    ).toBe("analysis-period");
+  });
+
+  it("returns stopped for a stopped holdout experiment", () => {
+    expect(getHoldoutStage({}, { status: "stopped" })).toBe("stopped");
+    expect(
+      getHoldoutStage({ analysisStartDate: new Date() }, { status: "stopped" }),
+    ).toBe("stopped");
+  });
+
+  describe("string-dates form", () => {
+    it("treats an ISO string analysisStartDate as the analysis period", () => {
+      expect(
+        getHoldoutStage(
+          { analysisStartDate: "2026-07-28T00:00:00.000Z" },
+          { status: "running" },
+        ),
+      ).toBe("analysis-period");
+    });
+
+    it("treats an absent analysisStartDate as running", () => {
+      expect(
+        getHoldoutStage(
+          { analysisStartDate: undefined },
+          { status: "running" },
+        ),
+      ).toBe("running");
+    });
+
+    it("treats a null analysisStartDate as running", () => {
+      expect(
+        getHoldoutStage({ analysisStartDate: null }, { status: "running" }),
+      ).toBe("running");
+    });
+
+    it("treats an empty-string analysisStartDate as running", () => {
+      expect(
+        getHoldoutStage({ analysisStartDate: "" }, { status: "running" }),
+      ).toBe("running");
+    });
   });
 });
